Day 2 Updates....

+++ OrderManagement.Core/DTO/OrderItemResponse.cs
@@ -8,8 +8,8 @@ public class OrderItemResponse
     public Guid OrderId { get; set; }
     public string? ProductName { get; set; }
     public int Quantity { get; set; }
-    public decimal UnitPrice { get; set; }
-    public decimal TotalPrice { get; set; }
+    public double UnitPrice { get; set; }
+    public double TotalPrice { get; set; }
 }
 
 public static class OrderItemExtensions
